chore: fix up pr comments
