+++ ios/Networking/Sources/Networking/URLSessionHTTPClient.swift
@@ -11,54 +11,39 @@ import Foundation
 
 @available(macOS 12.0, *)
 public struct URLSessionHTTPClient: HTTPClient {
-  private var urlSession: URLSessionProtocol
 
-  public init(urlSession: URLSessionProtocol) {
-    self.urlSession = urlSession
+  // MARK: Lifecycle
+
+  public init(urlSession: HTTPSession) {
+    session = urlSession
   }
+
+  // MARK: Public
 
   public func get(from url: URL) async -> HTTPClient.Result {
     do {
-      let (data, urlResponse) = try await urlSession.data(from: url)
+      let (data, urlResponse) = try await session.data(from: url)
       guard let httpUrlResponse = urlResponse as? HTTPURLResponse else {
         return .failure(HTTPClientError.invalidHTTPResponse)
       }
       return .success((data, httpUrlResponse))
-    } catch let networkError as URLError where networkError.code == .networkConnectionLost || networkError.code == .notConnectedToInternet {
+    } catch {
       return .failure(HTTPClientError.networkFailure)
-    } catch let error {
-      return .failure(error)
     }
   }
+
+  // MARK: Private
+
+  private var session: HTTPSession
 }
 
-// MARK: - URLSessionProtocol
+// MARK: - HTTPSession
 
-public protocol URLSessionProtocol {
+public protocol HTTPSession {
   func data(from url: URL) async throws -> (Data, URLResponse)
 }
 
-// MARK: - URLSession + URLSessionProtocol
+// MARK: - URLSession + HTTPSession
+
 @available(macOS 12.0, *)
-extension URLSession: URLSessionProtocol { }
-
-// MARK: - MockURLSession
-
-public struct MockURLSession: URLSessionProtocol {
-  public var data: Data
-  public var urlResponse: URLResponse
-  public var error: Error?
-
-  public init(data: Data, urlResponse: URLResponse, error: Error? = nil) {
-    self.data = data
-    self.urlResponse = urlResponse
-    self.error = error
-  }
-
-  public func data(from _: URL) async throws -> (Data, URLResponse) {
-    guard let error = error else {
-      return (data, urlResponse)
-    }
-    throw error
-  }
-}
+extension URLSession: HTTPSession { }
